requirements.txt
requests
beautifulsoup4
playwright
pandas
snowflake-connector-python[pandas]
python-dotenv